Implemented Moving, Deleting, Rotating, and Scaling Regions
Regions of pixels of artboards can be moved, deleted, rotated, and scaled, conforming to the rules of layers.

+++ src/cs/csss/project/io/ProjectSizeAndPositions.java
@@ -1,12 +1,12 @@
 package cs.csss.project.io;
 
 public record ProjectSizeAndPositions(
-	int leftmostX , 
-	int rightmostX , 
-	int lowermostY , 
-	int uppermostY , 
-	int width , 
-	int height , 
-	int midpointX , 
-	int midpointY
+	float leftmostX , 
+	float rightmostX , 
+	float lowermostY , 
+	float uppermostY , 
+	float width , 
+	float height , 
+	float midpointX , 
+	float midpointY
 ) {}
